Struct block_io::BlockBounds

pub struct BlockBounds {
    pub range: Range<usize>,
    pub first_block_offset: usize,
    pub last_block_offset: usize,
}

Block-wise bounds information for a data transfer (read or write) that has been calculated from a byte-wise offset and buffer length.

See the block_bounds() method.

Fields
range: Range<usize>

A Range from the first block to the last block of the transfer.

first_block_offset: usize

The offset into the first block (beginning bound) where the transfer starts.

last_block_offset: usize

The offset into the last block (ending bound) where the transfer ends.

Implementations

impl BlockBounds

pub fn block_count(&self) -> usize

The total number of blocks to be transferred, i.e., last block - first block.

pub fn is_first_block_aligned(&self) -> bool

Returns true if the first block of the transfer is aligned to a block boundary.

pub fn is_last_block_aligned(&self) -> bool

Returns true if the last block of the transfer is aligned to a block boundary.

pub fn block_bounds(
offset: usize,
length: usize,
block_size: &BlockSize
) -> Result<BlockBounds, &'static str>

Calculates block-wise bounds based on a byte-wise offset and buffer length for some block based device.

Arguments

offset: the absolute byte offset from the beginning of the block storage device at which the read/write starts.

length: the number of bytes to be read/written.

block_size: the relevant size information for the underlying block device.

Return

Returns a BlockBounds object.

If offset + length extends past the bounds, the Range will be truncated to the last block of the storage device (all the way to the end), and the last_block_offset will be 0.

Returns an error if the offset extends past the bounds of this block device.