Crate block_io

Wrappers for converting block I/O operations from one block size to another.

For example, these wrappers can expose a storage device that transfers 512-byte blocks at a time as a device that can transfer arbitrary bytes at a time (as little as one byte).

Limitations

Currently, the BlockIo struct is hardcoded to use a StorageDevice reference, when in reality it should just use anything that implements traits like BlockReader + BlockWriter.

The read and write functions are implemented such that if the backing storage device needs to be accessed, it is done so by transferring only one block at a time. This is quite inefficient, and we should instead transfer multiple blocks at once.

Cached blocks are stored as vectors of bytes on the heap, we should do something else such as separate mapped regions. Cached blocks cannot yet be dropped to relieve memory pressure.

Structs

BlockIo
A wrapper around a StorageDevice that supports reads and writes of arbitrary byte lengths (down to a single byte) by issuing commands to the underlying storage device. This is needed because most storage devices only allow reads/writes of larger blocks, e.g., a 512-byte sector or 4KB cluster.